IGNITE-24879 Sql. Add planner tests for temporal types coersion
Move shared functions to the base class.
