+++ modules/sql-engine/src/test/java/org/apache/ignite/internal/sql/engine/planner/datatypes/DateTimeMergeSourcesCoercionTest.java
@@ -1,27 +1,33 @@
-package org.apache.ignite.internal.sql.engine.planner.datatypes;
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
-import static org.hamcrest.MatcherAssert.assertThat;
+package org.apache.ignite.internal.sql.engine.planner.datatypes;
 
 import java.util.EnumMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.stream.Stream;
-import org.apache.calcite.rel.RelNode;
-import org.apache.calcite.rex.RexCall;
 import org.apache.calcite.rex.RexNode;
 import org.apache.ignite.internal.sql.engine.planner.datatypes.utils.DatetimePair;
 import org.apache.ignite.internal.sql.engine.planner.datatypes.utils.TypePair;
 import org.apache.ignite.internal.sql.engine.planner.datatypes.utils.Types;
-import org.apache.ignite.internal.sql.engine.rel.AbstractIgniteJoin;
-import org.apache.ignite.internal.sql.engine.rel.IgniteProject;
-import org.apache.ignite.internal.sql.engine.rel.IgniteRel;
-import org.apache.ignite.internal.sql.engine.rel.IgniteTableModify;
 import org.apache.ignite.internal.sql.engine.schema.IgniteSchema;
 import org.apache.ignite.internal.sql.engine.util.SqlTestUtils;
 import org.apache.ignite.internal.type.NativeTypes;
-import org.hamcrest.BaseMatcher;
-import org.hamcrest.Description;
 import org.hamcrest.Matcher;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -293,32 +299,5 @@ public class DateTimeMergeSourcesCoercionTest extends BaseTypeCoercionTest {
                 .opMatches(castTo(Types.TIMESTAMP_WLTZ_9)));
 
         return argsForMergeWithColumnAsValue().map(v -> diff.getOrDefault(v.get()[0], v));
-    }
-
-    private static Matcher<IgniteRel> mergeOperandMatcher(Matcher<RexNode> matcher) {
-        return new BaseMatcher<>() {
-            @Override
-            public boolean matches(Object actual) {
-                RelNode input = ((IgniteTableModify) actual).getInput();
-
-                List<RexNode> operands = null;
-                if (input instanceof IgniteProject) {
-                    operands = ((IgniteProject) input).getProjects();
-                } else if (input instanceof AbstractIgniteJoin) {
-                    RexCall condition = ((RexCall) ((AbstractIgniteJoin) input).getCondition());
-                    operands = condition.getOperands();
-                }
-
-                RexNode operand = Objects.requireNonNull(operands).get(1);
-                assertThat(operand, matcher);
-
-                return true;
-            }
-
-            @Override
-            public void describeTo(Description description) {
-
-            }
-        };
     }
 }
