+++ modules/sql-engine/src/test/java/org/apache/ignite/internal/sql/engine/planner/datatypes/NumericQuantifyOperatorTypeCoercionTest.java
@@ -17,32 +17,15 @@
 
 package org.apache.ignite.internal.sql.engine.planner.datatypes;
 
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.instanceOf;
-import static org.hamcrest.Matchers.not;
-
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
-import org.apache.calcite.rel.RelNode;
-import org.apache.calcite.rel.core.Join;
-import org.apache.calcite.rel.core.Project;
-import org.apache.calcite.rex.RexCall;
-import org.apache.calcite.rex.RexInputRef;
 import org.apache.calcite.rex.RexNode;
-import org.apache.calcite.rex.RexShuttle;
-import org.apache.calcite.sql.SqlKind;
-import org.apache.calcite.util.Util;
 import org.apache.ignite.internal.sql.engine.planner.datatypes.utils.NumericPair;
 import org.apache.ignite.internal.sql.engine.planner.datatypes.utils.TypePair;
 import org.apache.ignite.internal.sql.engine.planner.datatypes.utils.Types;
-import org.apache.ignite.internal.sql.engine.rel.IgniteRel;
 import org.apache.ignite.internal.sql.engine.schema.IgniteSchema;
 import org.apache.ignite.internal.type.NativeTypes;
-import org.hamcrest.BaseMatcher;
-import org.hamcrest.Description;
 import org.hamcrest.Matcher;
-import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
@@ -652,109 +635,5 @@ public class NumericQuantifyOperatorTypeCoercionTest extends BaseTypeCoercionTes
     @Test
     void argsIncludesAllTypePairs() {
         checkIncludesAllNumericTypePairs(args());
-    }
-
-    private static Matcher<IgniteRel> equalsToSomeOperatorMatcher(
-            Matcher<RexNode> first, Matcher<RexNode> second
-    ) {
-        return new BaseMatcher<>() {
-            @Override
-            public boolean matches(Object actual) {
-                RelNode projectRel = (RelNode) actual;
-                Join joinRel = (Join) projectRel.getInput(0);
-
-                RexNode comparison = joinRel.getCondition();
-
-                assertThat(comparison, instanceOf(RexCall.class));
-
-                RexCall comparisonCall = (RexCall) comparison;
-
-                RexNode leftOperand = comparisonCall.getOperands().get(0);
-                RexNode rightOperand = comparisonCall.getOperands().get(1);
-
-                assertThat(leftOperand, first);
-                assertThat(rightOperand, second);
-
-                return true;
-            }
-
-            @Override
-            public void describeTo(Description description) {
-
-            }
-        };
-    }
-
-    private static Matcher<IgniteRel> quantifyOperatorMatcher(
-            Matcher<RexNode> first, Matcher<RexNode> second, int firstOpIdx, int secondOpIdx
-    ) {
-        return new BaseMatcher<>() {
-            @Override
-            public boolean matches(Object actual) {
-                RexNode expression = ((Project) actual).getProjects().get(0);
-
-                List<RexNode> comparisonToVerify = findComparison(expression, firstOpIdx, secondOpIdx);
-
-                assertThat(comparisonToVerify, not(Matchers.empty()));
-
-                for (RexNode comparison : comparisonToVerify) {
-                    assertThat(comparison, instanceOf(RexCall.class));
-
-                    RexCall comparisonCall = (RexCall) comparison;
-
-                    RexNode leftOperand = comparisonCall.getOperands().get(0);
-                    RexNode rightOperand = comparisonCall.getOperands().get(1);
-
-                    assertThat(leftOperand, first);
-                    assertThat(rightOperand, second);
-                }
-
-                return true;
-            }
-
-            @Override
-            public void describeTo(Description description) {
-
-            }
-        };
-    }
-
-    /** Finds a comparison call comparing input references with provided indexes. */
-    private static List<RexNode> findComparison(RexNode expression, int firstOpIdx, int secondOpIdx) {
-        List<RexNode> result = new ArrayList<>();
-        new RexShuttle() {
-            @Override
-            public RexNode visitCall(RexCall call) {
-                if (call.getKind().belongsTo(SqlKind.BINARY_COMPARISON)) {
-                    result.add(call);
-
-                    return call;
-                }
-
-                return super.visitCall(call);
-            }
-        }.apply(expression);
-
-        result.removeIf(cmp -> {
-            try {
-                new RexShuttle() {
-                    @Override
-                    public RexNode visitInputRef(RexInputRef inputRef) {
-                        int column = inputRef.getIndex();
-                        if (column != firstOpIdx && column != secondOpIdx) {
-                            throw Util.FoundOne.NULL;
-                        }
-
-                        return inputRef;
-                    }
-                }.apply(cmp);
-            } catch (Util.FoundOne ignored) {
-                return true;
-            }
-
-            return false;
-        });
-
-        return result;
     }
 }
